--- pkg/flow/grpc-revisions.go
@@ -1,1 +1,0 @@
-package flow
